+++ CHANGELOG.md
@@ -25,11 +25,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3601](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3601))
 
 ### Fixed
-
 - `opentelemetry-instrumentation-asgi`: fix excluded_urls in instrumentation-asgi
   ([#3567](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3567))
 - `opentelemetry-resource-detector-containerid`: make it more quiet on platforms without cgroups
   ([#3579](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3579))
+- `opentelemetry-instrumentation-system-metrics`: Add process.runtime.cpython.gc_count metrics with collection unit is specified in semconv ([3617](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3617))
 
 ## Version 1.34.0/0.55b0 (2025-06-04)